move random func to util.js
