+++ src/mock/film.js
@@ -1,16 +1,4 @@
-
-const getRandomInteger = (min = 0, max = 1) => {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-const getRandomFloat = (min = 0.0, max = 1.0, numberDecimals = 1) => {
-  const randomValue = Math.random() * (max - min + 1) + min;
-
-  return Number(randomValue.toFixed(numberDecimals));
-};
+import { getRandomInteger, getRandomFloat } from '../util';
 
 const getRandomTitle = () => {
   const titles = [
